pyproject.toml
[project]
name = "fast-news-reader"
version = "0.1.0"
description = "Home Assistant custom integration: RSS reader with full image extraction (incl. content:encoded)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "fastender" }]
dependencies = [
  "feedparser>=6.0.10",
  "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-homeassistant-custom-component>=0.13",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
